Initial draft of virtualization support
This PR adds the initial draft of virtualization support by providing three new files: generate_pyrsa_images.sh, Dockerfile and Singularity. The first is a small bash script that runs neurodocker to create build files, either Docker or Singularity. The other files are the respective outcomes, that is a Dockerfile or Singularity recipe that can be used to build the images. This should allow to make and track changes in a reproducible manner.